pyproject.toml
[project]
name = "jailbreaking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "boto3>=1.42.58",
    "diffusers>=0.36.0",
    "editdistance>=0.8.1",
    "jiwer>=4.0.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "peft>=0.18.1",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "qwen-vl-utils>=0.0.14",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "sentencepiece>=0.2.1",
    "tabulate>=0.9.0",
    "tenacity>=9.1.2",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "transformers==5.5.0",
    "xai-sdk>=1.7.0",
]

[tool.uv.sources]
diffusers = { git = "https://github.com/huggingface/diffusers.git" }
